refactored YoutubeAPI out of utilities
