+++ stam.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import threading
 from servicemodules.discord import DiscordHandler
 from servicemodules.slack import SlackHandler
@@ -9,9 +8,6 @@ from structlog import get_logger
 from modules.module import Module
 from config import (
     database_path,
-    prod_local_path,
-    ENVIRONMENT_TYPE,
-    acceptable_environment_types,
 )
 from servicemodules.serviceConstants import Services
 
@@ -23,7 +19,9 @@ def get_stampy_modules() -> dict[str, Module]:
     """Dynamically import and return all Stampy modules"""
     stampy_modules = {}
     path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "modules")
-    for file_title in [f[:-3] for f in os.listdir(path) if f.endswith(".py") and f != "__init__.py"]:
+    for file_title in [
+        f[:-3] for f in os.listdir(path) if f.endswith(".py") and f != "__init__.py"
+    ]:
         log.info("import", filename=file_title)
         mod = __import__(".".join(["modules", file_title]), fromlist=[file_title])
         log.info("import", module_name=mod)
@@ -40,7 +38,7 @@ if __name__ == "__main__":
     utils = Utilities.get_instance()
 
     if not os.path.exists(database_path):
-        raise Exception("Couldn't find the stampy database file at " + f"{database_path}")
+        raise Exception(f"Couldn't find the stampy database file at {database_path}")
 
     utils.modules_dict = get_stampy_modules()
 
@@ -53,10 +51,10 @@ if __name__ == "__main__":
     service_threads = []
     e = threading.Event()
     utils.stop = e
-    for service in utils.service_modules_dict:
-        log.info(log_type, msg=f"Starting {service}")
-        service_threads.append(utils.service_modules_dict[service].start(e))
-        log.info(log_type, msg=f"{service} Started!")
+    for service_name, service in utils.service_modules_dict.items():
+        log.info(log_type, msg=f"Starting {service_name}")
+        service_threads.append(service.start(e))
+        log.info(log_type, msg=f"{service_name} Started!")
 
     for thread in service_threads:
         if thread.is_alive() and not thread.daemon:
